Refactor client/src/document*.* files
Fixes #412
