+++ client/src/document/ingredients/browser-compatibility-table/bcd.scss
@@ -144,14 +144,14 @@ $link-color: #3d7e9a;
 }
 
 .bc-table .bc-platforms .bc-platform-desktop:before {
-  background-image: url("../../icons/platforms/desktop.svg");
+  background-image: url("./icons/platforms/desktop.svg");
   background-size: 24px;
   width: 24px;
   height: 24px;
 }
 
 .bc-table .bc-platforms .bc-platform-mobile:before {
-  background-image: url("../../icons/platforms/mobile.svg");
+  background-image: url("./icons/platforms/mobile.svg");
   background-size: 16px;
   width: 16px;
   height: 32px;
@@ -159,7 +159,7 @@ $link-color: #3d7e9a;
 }
 
 .bc-table .bc-platforms .bc-platform-server:before {
-  background-image: url("../../icons/platforms/server.svg");
+  background-image: url("./icons/platforms/server.svg");
   background-size: 24px;
   width: 24px;
   height: 24px;
@@ -249,7 +249,7 @@ $link-color: #3d7e9a;
 
 .bc-browsers .bc-head-icon-chrome:before,
 .bc-browsers .bc-head-icon-chrome_android:before {
-  background-image: url("../../icons/browsers/chrome.svg");
+  background-image: url("./icons/browsers/chrome.svg");
   background-size: 20px;
   width: 20px;
   height: 24px;
@@ -257,7 +257,7 @@ $link-color: #3d7e9a;
 
 .bc-browsers .bc-head-icon-edge:before,
 .bc-browsers .bc-head-icon-edge_mobile:before {
-  background-image: url("../../icons/browsers/edge.svg");
+  background-image: url("./icons/browsers/edge.svg");
   background-size: 20px;
   width: 20px;
   height: 24px;
@@ -265,14 +265,14 @@ $link-color: #3d7e9a;
 
 .bc-browsers .bc-head-icon-firefox:before,
 .bc-browsers .bc-head-icon-firefox_android:before {
-  background-image: url("../../icons/browsers/firefox.svg");
+  background-image: url("./icons/browsers/firefox.svg");
   background-size: 20px;
   width: 20px;
   height: 24px;
 }
 
 .bc-browsers .bc-head-icon-ie:before {
-  background-image: url("../../icons/browsers/internet-explorer.svg");
+  background-image: url("./icons/browsers/internet-explorer.svg");
   background-size: 20px;
   width: 20px;
   height: 24px;
@@ -280,7 +280,7 @@ $link-color: #3d7e9a;
 
 .bc-browsers .bc-head-icon-opera:before,
 .bc-browsers .bc-head-icon-opera_android:before {
-  background-image: url("../../icons/browsers/opera.svg");
+  background-image: url("./icons/browsers/opera.svg");
   background-size: 20px;
   width: 20px;
   height: 24px;
@@ -288,14 +288,14 @@ $link-color: #3d7e9a;
 
 .bc-browsers .bc-head-icon-safari:before,
 .bc-browsers .bc-head-icon-safari_ios:before {
-  background-image: url("../../icons/browsers/safari.svg");
+  background-image: url("./icons/browsers/safari.svg");
   background-size: 20px;
   width: 20px;
   height: 24px;
 }
 
 .bc-browsers .bc-head-icon-webview_android:before {
-  background-image: url("../../icons/platforms/android.svg");
+  background-image: url("./icons/platforms/android.svg");
   background-size: 18px;
   width: 18px;
   height: 22px;
@@ -303,14 +303,14 @@ $link-color: #3d7e9a;
 }
 
 .bc-browsers .bc-head-icon-samsunginternet_android:before {
-  background-image: url("../../icons/browsers/samsung-internet.svg");
+  background-image: url("./icons/browsers/samsung-internet.svg");
   background-size: 20px;
   width: 20px;
   height: 24px;
 }
 
 .bc-browsers .bc-head-icon-nodejs:before {
-  background-image: url("../../icons/platforms/nodejs.svg");
+  background-image: url("./icons/platforms/nodejs.svg");
   background-size: 20px;
   width: 20px;
   height: 24px;
@@ -481,14 +481,14 @@ td[tabindex]:hover .bc-history-link {
 }
 
 .ic-footnote:before {
-  background-image: url("../../icons/general/asterisk.svg");
+  background-image: url("../../../icons/general/asterisk.svg");
   background-size: 10px;
   width: 10px;
   height: 10px;
 }
 
 .ic-history:before {
-  background-image: url("../../icons/arrows/caret-down.svg");
+  background-image: url("../../../icons/arrows/caret-down.svg");
   background-size: 9px;
   width: 9px;
   height: 13px;
@@ -500,25 +500,25 @@ td[tabindex].bc-supports-no:hover .ic-history:before,
 .active.bc-supports-yes .ic-history:before,
 td[tabindex].bc-supports-yes:focus .ic-history:before,
 td[tabindex].bc-supports-yes:hover .ic-history:before {
-  background-image: url("../../icons/bcd/no-support.svg");
+  background-image: url("./icons/general/no-support.svg");
 }
 
 .ic-deprecated:before {
-  background-image: url("../../icons/emojis/thumbs-down.svg");
+  background-image: url("../../../icons/emojis/thumbs-down.svg");
   background-size: 16px;
   width: 16px;
   height: 20px;
 }
 
 .ic-non-standard:before {
-  background-image: url("../../icons/bcd/exclamation-triangle.svg");
+  background-image: url("./icons/general/exclamation-triangle.svg");
   background-size: 17px;
   width: 17px;
   height: 20px;
 }
 
 .ic-experimental:before {
-  background-image: url("../../icons/bcd/flask.svg");
+  background-image: url("./icons/general/flask.svg");
   background-size: 16px;
   width: 16px;
   height: 20px;
@@ -535,14 +535,14 @@ td[tabindex].bc-supports-yes:hover .ic-history:before {
 }
 
 .ic-disabled:before {
-  background-image: url("../../icons/bcd/flag.svg");
+  background-image: url("./icons/general/flag.svg");
   background-size: 24px;
   width: 24px;
   height: 24px;
 }
 
 .ic-altname:before {
-  background-image: url("../../icons/bcd/alternate.svg");
+  background-image: url("./icons/general/alternate.svg");
   background-size: 12px;
   width: 12px;
   height: 22px;
