+++ client/src/document/ingredients/browser-compatibility-table/browser-support-detail.js
