+++ client/src/document/ingredients/browser-compatibility-table/browsers.js
