+++ client/src/document/ingredients/browser-compatibility-table/error-boundary.js
