+++ client/src/document/ingredients/browser-compatibility-table/index.js
