+++ client/src/document/ingredients/browser-compatibility-table/legend.js
