+++ client/src/document/ingredients/browser-compatibility-table/platforms.js
