+++ client/src/document/ingredients/browser-compatibility-table/utils/browser-name.js
