+++ client/src/document/ingredients/interactive-example.js
