+++ client/src/document/ingredients/specifications.js
